Clean up GEM trigger modules
