+++ L1Trigger/L1TGEM/plugins/GEMPadDigiClusterProducer.cc
@@ -13,6 +13,8 @@
  *  \author Sven Dildick (TAMU)
  */
 
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/PluginManager/interface/ModuleDef.h"
 #include "FWCore/Framework/interface/stream/EDProducer.h"
 #include "FWCore/Framework/interface/MakerMacros.h"
 #include "FWCore/Framework/interface/ESHandle.h"
