+++ L1Trigger/L1TGEM/plugins/GEMPadDigiProducer.cc
@@ -2,6 +2,7 @@
 #define L1Trigger_L1TGEM_GEMPadDigiProducer_h
 
 #include "FWCore/Framework/interface/stream/EDProducer.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
 #include "FWCore/Framework/interface/ESHandle.h"
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/EventSetup.h"
@@ -21,12 +22,10 @@
 #include <set>
 
 /// \class GEMPadDigiProducer
-/// producer for GEM-CSC trigger pads
+/// producer for GEM trigger pads
 
 class GEMPadDigiProducer : public edm::stream::EDProducer<> {
 public:
-  //typedef GEMDigitizer::StripDigiSimLinks StripDigiSimLinks;
-
   explicit GEMPadDigiProducer(const edm::ParameterSet& ps);
 
   ~GEMPadDigiProducer() override;
@@ -108,5 +107,7 @@ void GEMPadDigiProducer::buildPads(const GEMDigiCollection& det_digis, GEMPadDig
     }
   }
 }
+
+DEFINE_FWK_MODULE(GEMPadDigiProducer);
 
 #endif
