+++ L1Trigger/L1TGEM/plugins/ME0PadDigiProducer.cc
@@ -2,6 +2,7 @@
 #define L1Trigger_L1TGEM_ME0PadDigiProducer_h
 
 #include "FWCore/Framework/interface/stream/EDProducer.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
 #include "FWCore/Framework/interface/ESHandle.h"
 #include "FWCore/Framework/interface/ConsumesCollector.h"
 #include "FWCore/Framework/interface/Event.h"
@@ -92,5 +93,7 @@ void ME0PadDigiProducer::buildPads(const ME0DigiCollection& det_digis, ME0PadDig
     }
   }
 }
+
+DEFINE_FWK_MODULE(ME0PadDigiProducer);
 
 #endif
