+++ L1Trigger/L1TGEM/plugins/ME0TriggerProducer.cc
@@ -1,11 +1,47 @@
-#include "L1Trigger/L1TGEM/plugins/ME0TriggerProducer.h"
-#include "L1Trigger/L1TGEM/interface/ME0TriggerBuilder.h"
-#include "DataFormats/Common/interface/Handle.h"
+#ifndef L1Trigger_L1TGEM_ME0TriggerProducer_h
+#define L1Trigger_L1TGEM_ME0TriggerProducer_h
+
+/** \class ME0TriggerProducer
+ *
+ * Takes ME0 pad clusters as input
+ * Produces ME0 trigger objects
+ *
+ * \author Sven Dildick (TAMU).
+ *
+ */
+
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/Framework/interface/ConsumesCollector.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/global/EDProducer.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/Utilities/interface/InputTag.h"
 #include "FWCore/Framework/interface/ESHandle.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "L1Trigger/L1TGEM/interface/ME0TriggerBuilder.h"
+#include "DataFormats/Common/interface/Handle.h"
+#include "DataFormats/GEMDigi/interface/ME0PadDigiCollection.h"
 #include "DataFormats/GEMDigi/interface/ME0TriggerDigiCollection.h"
 #include "Geometry/Records/interface/MuonGeometryRecord.h"
 #include "Geometry/GEMGeometry/interface/ME0Geometry.h"
+
+class ME0TriggerBuilder;
+
+class ME0TriggerProducer : public edm::global::EDProducer<> {
+public:
+  explicit ME0TriggerProducer(const edm::ParameterSet&);
+  ~ME0TriggerProducer() override;
+
+  //virtual void beginRun(const edm::EventSetup& setup);
+  void produce(edm::StreamID, edm::Event&, const edm::EventSetup&) const override;
+
+private:
+  edm::InputTag me0PadDigis_;
+  edm::EDGetTokenT<ME0PadDigiCollection> me0_pad_token_;
+  edm::ParameterSet config_;
+  bool useClusters_;
+};
 
 ME0TriggerProducer::ME0TriggerProducer(const edm::ParameterSet& conf) {
   me0PadDigis_ = conf.getParameter<edm::InputTag>("ME0PadDigis");
@@ -38,3 +74,7 @@ void ME0TriggerProducer::produce(edm::StreamID, edm::Event& ev, const edm::Event
   // Put collections in event.
   ev.put(std::move(oc_trig));
 }
+
+DEFINE_FWK_MODULE(ME0TriggerProducer);
+
+#endif
