--- L1Trigger/L1TGEM/plugins/ME0TriggerProducer.h
@@ -1,38 +1,0 @@
-#ifndef L1Trigger_L1TGEM_ME0TriggerProducer_h
-#define L1Trigger_L1TGEM_ME0TriggerProducer_h
-
-/** \class ME0TriggerProducer
- *
- * Takes ME0 pad clusters as input
- * Produces ME0 trigger objects
- *
- * \author Sven Dildick (TAMU).
- *
- */
-
-#include "DataFormats/GEMDigi/interface/ME0PadDigiCollection.h"
-#include "FWCore/Framework/interface/ConsumesCollector.h"
-#include "FWCore/Framework/interface/Frameworkfwd.h"
-#include "FWCore/Framework/interface/global/EDProducer.h"
-#include "FWCore/Framework/interface/Event.h"
-#include "FWCore/ParameterSet/interface/ParameterSet.h"
-#include "FWCore/Utilities/interface/InputTag.h"
-
-class ME0TriggerBuilder;
-
-class ME0TriggerProducer : public edm::global::EDProducer<> {
-public:
-  explicit ME0TriggerProducer(const edm::ParameterSet&);
-  ~ME0TriggerProducer() override;
-
-  //virtual void beginRun(const edm::EventSetup& setup);
-  void produce(edm::StreamID, edm::Event&, const edm::EventSetup&) const override;
-
-private:
-  edm::InputTag me0PadDigis_;
-  edm::EDGetTokenT<ME0PadDigiCollection> me0_pad_token_;
-  edm::ParameterSet config_;
-  bool useClusters_;
-};
-
-#endif
